add compose efun
This required updating the CallEfun instruction to use a more globally-accessible collection to index into for the name, to allow us to use a hand-written assembly function(!) to support the efun.

The reason we use a hand-written function, rather than just using another efun, is that it's much easier to leverage the vm to handle things like resource limits, rather than just executing everything directly.

The compromise in all of this is that the EFUN_PROTOTYPES map can no longer be reordered without breaking a bunch of tests, and invaliding any persisted code.