implement 'synchronized' keyword
due to the multi-threaded implementation of the VM, there needs to be a mechanism to avoid threads overwriting each other while working. The initial step is going to be allowing functions to be flagged as synchronized. When a synchronized function is run, it locks the entire object that the function belongs to, for the exclusive use of the Task that requested the lock. That task has unfettered access to the object, including being able to call other synchronized methods without needing to re-take the lock.

The lock is taken prior to setting up arguments for the call.

Deadlocks are handled by the Task timeout. They will be killed and a runtime error emitted in those cases.